Peer operations

+++ wireguard_server.sh
@@ -1,25 +1,8 @@
 # !/bin/sh
-# cat > ${client_config} <<EOF && chmod 600 ${client_config}
-# [Interface]
-# Address = ${client_wg_ip}/${cidr}
-# PrivateKey = $(head -1 ${client_private_key})
-# DNS = $(echo ${client_dns_ips} | sed -E 's/ +/, /g')
-# [Peer]
-# PublicKey = $(head -1 ${server_public_key})
-# PresharedKey = $(head -1 ${preshared_key})
-# AllowedIPs = $(echo ${client_allowed_ips} | sed -E 's/ +/, /g')
-# Endpoint = ${server_public_ip}:${server_port}
-# PersistentKeepalive = 25
-# EOF
-
-#     cat >> "/etc/wireguard/${interface_name}.conf" <<EOF
-# PostUp = iptables -A FORWARD -i %i -j ACCEPT; iptables -A FORWARD -o %i -j ACCEPT; iptables -t nat -A POSTROUTING -o ${forwarding_interface} -j MASQUERADE
-# PostDown = iptables -D FORWARD -i %i -j ACCEPT; iptables -D FORWARD -o %i -j ACCEPT; iptables -t nat -D POSTROUTING -o ${forwarding_interface} -j MASQUERADE
-# EOF
 
 help=false
 visualize=false
-syntax=false # While false it hasn't found any command
+syntax=false # While false it hasn't found any command on the execution thread
 
 check_run(){ # Run configuration commands
     if [ $# -gt 1 ]; then
@@ -32,6 +15,14 @@ check_run(){ # Run configuration commands
 reminder(){
     echo "Use wg-quick up/down {<interface name>} to enable or disable an interface without deleting it"
     echo "Use wg show [<interface name>] to view the status of current interfaces"
+    echo "Check wireguard service status: systemctl status wg-quick@{<interface-name>}"
+}
+show(){
+    if [ $# -gt 0 ]; then
+        check_run "cat /etc/wireguard/${1}.conf"
+    else
+        check_run "ls /etc/wireguard/"
+    fi
 }
 new(){
     # Mandatory args
@@ -70,6 +61,7 @@ new(){
     fi
 
     check_run "wg-quick up ${interface_name}"
+    echo "To autostart this interface execute: sudo systemctl enable wg-quick@${interface_name}.service"
 }
 enable-forwarding(){
     # Mandatory args
@@ -79,6 +71,11 @@ enable-forwarding(){
     fi
     interface_name=$1
     forwarding_interface=$2
+
+    if [ ! -f /etc/wireguard/${interface_name}.conf ]; then
+        echo "Interface config file not found"
+        return
+    fi
 
     file=""
     forwarded=false
@@ -100,7 +97,7 @@ enable-forwarding(){
     fi
 
     file="[Interface]\nPostUp = iptables -A FORWARD -i %i -j ACCEPT; iptables -A FORWARD -o %i -j ACCEPT; iptables -t nat -A POSTROUTING -o ${forwarding_interface} -j MASQUERADE\nPostDown = iptables -D FORWARD -i %i -j ACCEPT; iptables -D FORWARD -o %i -j ACCEPT; iptables -t nat -D POSTROUTING -o ${forwarding_interface} -j MASQUERADE\n"${file}
-    echo "Adding PostUp and Post down rules to ${interface_name}.config"
+    echo "Adding PostUp and Post down rules to ${interface_name}.conf"
     printf "$file" > /etc/wireguard/${interface_name}.conf
     echo "Allowing traffic forwarding at system level"
     sysctl -w net.ipv4.ip_forward=1
@@ -112,6 +109,11 @@ disable-forwarding(){
         return
     fi
     interface_name=$1
+
+    if [ ! -f /etc/wireguard/${interface_name}.conf ]; then
+        echo "Interface config file not found"
+        return
+    fi
 
     file=""
     forwarded=false
@@ -126,9 +128,10 @@ disable-forwarding(){
         fi
     done < /etc/wireguard/${interface_name}.conf;
 
+
     if [ $forwarded == "true" ]; then
         printf "$file" > /etc/wireguard/${interface_name}.conf
-        echo "Removed PostUp and Post down rules from ${interface_name}.config"
+        echo "Removed PostUp and Post down rules from ${interface_name}.conf"
     else
         echo "Could not find redirection rules in the config file"
     fi
@@ -141,15 +144,86 @@ remove(){
         return
     fi
     interface_name=$1
-    read -p "Confirm by retyping the interface name: " confirmation
+    if [ ! -f /etc/wireguard/${interface_name}.conf ]; then
+        echo "Interface config file not found"
+        return
+    fi
+
+    read -p "Retype the interface name to confirm: " confirmation
     if [ "$confirmation" != "$interface_name" ]; then echo "Aborted." && return; fi
 
+
     check_run "wg-quick down ${interface_name}"
+    check_run "systemctl stop wg-quick@${interface_name}.service"
+    check_run "systemctl disable wg-quick@${interface_name}.service"
     check_run "rm /etc/wireguard/${interface_name}.conf"
+    check_run "systemctl daemon-reload"
+    check_run "systemctl reset-failed"
     echo "To completely remove the VPN configuration please close the port by removing the UFW rule: 'ufw status' && 'ufw delete RULE'"
 }
 peer-new(){
-    echo "Not implemented"
+    # Mandatory args
+    if [ ! $# -ge 1 ] || [[ "$1" == "-"* ]] ; then
+        help=true
+        return
+    fi
+    interface_name=$1
+    shift
+    # Labelled args
+    public_key=false
+    ip=false
+    name=""
+    while [ $# -gt 0 ] ; do
+        case "$1" in
+            -f) shift;
+                if [ -f $1 ]; then
+                    public_key=$(<$1)
+                else
+                    echo "Public key file does not exist"; return
+                fi
+            ;;
+            -k) shift; public_key=$1;;
+            -i) shift; ip=$1;;
+            -n) shift; name="# $1 \n";;
+        esac
+        shift
+    done
+
+    if [ ! -f /etc/wireguard/${interface_name}.conf ]; then
+        echo "Interface config file not found"
+        return
+    fi
+
+    if [ "$public_key" == "false" ] || [ "$ip" = "false" ]; then
+        help=true
+        return
+    fi
+
+    if [[ "${ip}" != *"/"* ]]; then
+        echo "Remember to suffix the IP such as 10.0.0.2/32"
+        return
+    fi
+
+    # Commands
+    peer="[Peer]\n${name}PublicKey = ${public_key}\nAllowedIPs = ${ip}\n"
+
+    while read line; do
+        if [ "$line" != "" ]; then
+            if [[ "$line" == *"PublicKey = ${public_key}"* ]]; then
+                echo "Aborted. Public key matches with that of another peer."
+                return
+            fi
+            if [[ "$line" == "AllowedIPs"* ]]; then
+                if [[ "$line" == *"${ip}"* ]]; then
+                    echo "Aborted. IP is not unique to this peer."
+                    return
+                fi
+            fi
+        fi
+    done < /etc/wireguard/${interface_name}.conf;
+
+    printf "\n${peer}" >> /etc/wireguard/${interface_name}.conf;
+    echo "Peer added to ${interface_name}.conf"
 }
 peer-enable(){
     echo "Not implemented"
@@ -158,7 +232,72 @@ peer-disable(){
     echo "Not implemented"
 }
 peer-remove(){
-    echo "Not implemented"
+    # Mandatory args
+    if [ ! $# -ge 1 ] || [[ "$1" == "-"* ]] ; then
+        help=true
+        return
+    fi
+    interface_name=$1
+    shift
+    ip=""
+    name=""
+    while [ $# -gt 0 ] ; do
+        case "$1" in
+            -n) shift; name=$1;;
+            -i) shift; ip=$1;;
+        esac
+        shift
+    done
+    if [ "$ip" == "" ] && [ "$name" == "" ]; then
+        help=true
+        return
+    fi
+    if [ ! -f /etc/wireguard/${interface_name}.conf ]; then
+        echo "Interface config file not found"
+        return
+    fi
+
+    file=""
+    peer_temp=""
+    peer_temp_delete=false
+    peer_deleted=false
+    peer_count=0
+    forwarded=false
+
+    while read line; do
+        if [ "$line" != "" ]; then
+            if [ "$peer_temp" == "" ] && [[ "$line" != "[Peer]"* ]]; then #pre-peer line
+                file="${file}${line}\n"
+            else # peer line
+                if [[ "$line" == "[Peer]"* ]]; then # new peer
+                    # Reset
+                    file="${file}${peer_temp}\n"
+                    peer_temp_delete=false
+                    peer_temp=""
+                    peer_count=$peer_count+1
+                # \/ check if peer meets deletion criteria
+                elif [[ "$line" == "AllowedIPs"* ]] && [[ "$line" == *"${ip}"* ]] || [ "$line" == "# ${name}" ]; then
+                    peer_temp_delete=true
+                    # \/ check if a different peer has been flagged for deletion
+                    if [ $peer_deleted != "false" ] && [ "$peer_count" != "$peer_deleted" ]; then
+                        echo "Aborted. Multiple peers meet the same criteria"
+                        return
+                    else
+                        peer_deleted=$peer_count
+                    fi
+                    peer_temp="" # clear peer
+                fi
+
+                if [ "$peer_temp_delete" == "false" ]; then # The peer is safe
+                    peer_temp="${peer_temp}${line}\n"
+                fi
+            fi
+        fi
+    done < /etc/wireguard/${interface_name}.conf;
+    file="${file}${peer_temp}\n"
+
+    printf "$file"
+
 }
